Fn 359 user exercise instance delete

+++ src/main/java/com/thefitnation/service/UserExerciseInstanceSetService.java
@@ -1,19 +1,15 @@
 package com.thefitnation.service;
 
-import com.thefitnation.domain.ExerciseInstanceSet;
-import com.thefitnation.domain.UserExerciseInstance;
-import com.thefitnation.domain.UserExerciseInstanceSet;
-import com.thefitnation.repository.ExerciseInstanceSetRepository;
-import com.thefitnation.repository.UserExerciseInstanceRepository;
-import com.thefitnation.repository.UserExerciseInstanceSetRepository;
-import com.thefitnation.service.dto.UserExerciseInstanceSetDTO;
-import com.thefitnation.service.mapper.UserExerciseInstanceSetMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.stereotype.Service;
+import com.thefitnation.domain.*;
+import com.thefitnation.repository.*;
+import com.thefitnation.security.*;
+import com.thefitnation.service.dto.*;
+import com.thefitnation.service.mapper.*;
+import java.util.*;
+import org.slf4j.*;
+import org.springframework.data.domain.*;
+import org.springframework.stereotype.*;
+import org.springframework.transaction.annotation.*;
 
 /**
  * Service Implementation for managing UserExerciseInstanceSet.
@@ -25,18 +21,17 @@ public class UserExerciseInstanceSetService {
     private final Logger log = LoggerFactory.getLogger(UserExerciseInstanceSetService.class);
 
     private final UserExerciseInstanceSetRepository userExerciseInstanceSetRepository;
-
     private final ExerciseInstanceSetRepository exerciseInstanceSetRepository;
-
     private final UserExerciseInstanceRepository userExerciseInstanceRepository;
-
     private final UserExerciseInstanceSetMapper userExerciseInstanceSetMapper;
+    private final UserRepository userRepository;
 
-    public UserExerciseInstanceSetService(UserExerciseInstanceSetRepository userExerciseInstanceSetRepository, ExerciseInstanceSetRepository exerciseInstanceSetRepository, UserExerciseInstanceRepository userExerciseInstanceRepository, UserExerciseInstanceSetMapper userExerciseInstanceSetMapper) {
+    public UserExerciseInstanceSetService(UserExerciseInstanceSetRepository userExerciseInstanceSetRepository, ExerciseInstanceSetRepository exerciseInstanceSetRepository, UserExerciseInstanceRepository userExerciseInstanceRepository, UserExerciseInstanceSetMapper userExerciseInstanceSetMapper, UserRepository userRepository) {
         this.userExerciseInstanceSetRepository = userExerciseInstanceSetRepository;
         this.exerciseInstanceSetRepository = exerciseInstanceSetRepository;
         this.userExerciseInstanceRepository = userExerciseInstanceRepository;
         this.userExerciseInstanceSetMapper = userExerciseInstanceSetMapper;
+        this.userRepository = userRepository;
     }
 
     /**
@@ -47,23 +42,14 @@ public class UserExerciseInstanceSetService {
      */
     public UserExerciseInstanceSetDTO save(UserExerciseInstanceSetDTO userExerciseInstanceSetDTO) {
         log.debug("Request to save UserExerciseInstanceSet : {}", userExerciseInstanceSetDTO);
+        Optional<User> user = userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin());
+        if (!user.isPresent()) {
+            return null;
+        }
         UserExerciseInstanceSet userExerciseInstanceSet = userExerciseInstanceSetMapper.userExerciseInstanceSetDTOToUserExerciseInstanceSet(userExerciseInstanceSetDTO);
         userExerciseInstanceSet = userExerciseInstanceSetRepository.save(userExerciseInstanceSet);
         addUserExerciseInstanceSetToParent(userExerciseInstanceSet);
-        UserExerciseInstanceSetDTO result = userExerciseInstanceSetMapper.userExerciseInstanceSetToUserExerciseInstanceSetDTO(userExerciseInstanceSet);
-        return result;
-    }
-
-    public void addUserExerciseInstanceSetToParent(UserExerciseInstanceSet userExerciseInstanceSet) {
-        UserExerciseInstance userExerciseInstance = userExerciseInstanceRepository.findOne((userExerciseInstanceSet.getUserExerciseInstance()).getId());
-        userExerciseInstance.addUserExerciseInstanceSet(userExerciseInstanceSet);
-        userExerciseInstanceRepository.save(userExerciseInstance);
-
-        if (userExerciseInstanceSet.getExerciseInstanceSet() != null) {
-            ExerciseInstanceSet exerciseInstanceSet = exerciseInstanceSetRepository.findOne((userExerciseInstanceSet.getExerciseInstanceSet()).getId());
-            exerciseInstanceSet.addUserExerciseInstanceSet(userExerciseInstanceSet);
-            exerciseInstanceSetRepository.save(exerciseInstanceSet);
-        }
+        return userExerciseInstanceSetMapper.userExerciseInstanceSetToUserExerciseInstanceSetDTO(userExerciseInstanceSet);
     }
 
     /**
@@ -75,8 +61,8 @@ public class UserExerciseInstanceSetService {
     @Transactional(readOnly = true)
     public Page<UserExerciseInstanceSetDTO> findAll(Pageable pageable) {
         log.debug("Request to get all UserExerciseInstanceSets");
-        Page<UserExerciseInstanceSet> result = userExerciseInstanceSetRepository.findAll(pageable);
-        return result.map(userExerciseInstanceSet -> userExerciseInstanceSetMapper.userExerciseInstanceSetToUserExerciseInstanceSetDTO(userExerciseInstanceSet));
+        Page<UserExerciseInstanceSet> result = userExerciseInstanceSetRepository.findAll(SecurityUtils.getCurrentUserLogin(), pageable);
+        return result.map(userExerciseInstanceSetMapper::userExerciseInstanceSetToUserExerciseInstanceSetDTO);
     }
 
     /**
@@ -88,9 +74,8 @@ public class UserExerciseInstanceSetService {
     @Transactional(readOnly = true)
     public UserExerciseInstanceSetDTO findOne(Long id) {
         log.debug("Request to get UserExerciseInstanceSet : {}", id);
-        UserExerciseInstanceSet userExerciseInstanceSet = userExerciseInstanceSetRepository.findOne(id);
-        UserExerciseInstanceSetDTO userExerciseInstanceSetDTO = userExerciseInstanceSetMapper.userExerciseInstanceSetToUserExerciseInstanceSetDTO(userExerciseInstanceSet);
-        return userExerciseInstanceSetDTO;
+        UserExerciseInstanceSet userExerciseInstanceSet = userExerciseInstanceSetRepository.findOne(SecurityUtils.getCurrentUserLogin(), id);
+        return userExerciseInstanceSetMapper.userExerciseInstanceSetToUserExerciseInstanceSetDTO(userExerciseInstanceSet);
     }
 
     /**
@@ -100,11 +85,13 @@ public class UserExerciseInstanceSetService {
      */
     public void delete(Long id) {
         log.debug("Request to delete UserExerciseInstanceSet : {}", id);
-        removeUserExerciseInstanceSetFromRelatedItems(id);
-        userExerciseInstanceSetRepository.delete(id);
+        if (userExerciseInstanceSetRepository.findOne(SecurityUtils.getCurrentUserLogin(), id) != null) {
+            removeUserExerciseInstanceSetFromRelatedItems(id);
+            userExerciseInstanceSetRepository.delete(id);
+        }
     }
 
-    public void removeUserExerciseInstanceSetFromRelatedItems(Long id) {
+    private void removeUserExerciseInstanceSetFromRelatedItems(Long id) {
         UserExerciseInstanceSet userExerciseInstanceSet = userExerciseInstanceSetRepository.findOne(id);
         if (userExerciseInstanceSet != null) {
             log.debug("Request to remove UserExerciseInstanceSet from UserExerciseInstance : {}", userExerciseInstanceSet.getId());
@@ -117,6 +104,18 @@ public class UserExerciseInstanceSetService {
                 exerciseInstanceSet.removeUserExerciseInstanceSet(userExerciseInstanceSet);
                 exerciseInstanceSetRepository.save(exerciseInstanceSet);
             }
+        }
+    }
+
+    private void addUserExerciseInstanceSetToParent(UserExerciseInstanceSet userExerciseInstanceSet) {
+        UserExerciseInstance userExerciseInstance = userExerciseInstanceRepository.findOne((userExerciseInstanceSet.getUserExerciseInstance()).getId());
+        userExerciseInstance.addUserExerciseInstanceSet(userExerciseInstanceSet);
+        userExerciseInstanceRepository.save(userExerciseInstance);
+
+        if (userExerciseInstanceSet.getExerciseInstanceSet() != null) {
+            ExerciseInstanceSet exerciseInstanceSet = exerciseInstanceSetRepository.findOne((userExerciseInstanceSet.getExerciseInstanceSet()).getId());
+            exerciseInstanceSet.addUserExerciseInstanceSet(userExerciseInstanceSet);
+            exerciseInstanceSetRepository.save(exerciseInstanceSet);
         }
     }
 }
